% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kebir_gb.R
\docType{data}
\name{kebir_gb}
\alias{kebir_gb}
\title{Kebir glioblastoma example dataset}
\format{
A Biobase ExpressionSet with gene-level expression values in
\code{Biobase::exprs(kebir_gb)} and sample annotations in
\code{Biobase::pData(kebir_gb)}.
}
\source{
Kebir et al.; GSE145128.
}
\usage{
kebir_gb
}
\description{
A small glioblastoma example dataset stored as a Biobase ExpressionSet.
It contains gene-level expression values and sample-level metadata from
Kebir et al. and can be used for package examples and tutorials.
}
\details{
Expression values can be accessed with \code{Biobase::exprs(kebir_gb)}.
Sample metadata can be accessed with \code{Biobase::pData(kebir_gb)}.
}
\examples{
if (requireNamespace("Biobase", quietly = TRUE)) {
  data(kebir_gb)

  expr_mat <- Biobase::exprs(kebir_gb)
  sample_info <- Biobase::pData(kebir_gb)

  dim(expr_mat)
  head(sample_info)
}
}
\keyword{datasets}
