% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmpaR_package.R
\docType{package}
\name{vmpaR-package}
\alias{vmpaR}
\alias{vmpaR-package}
\title{vmpaR: context-matched protein activity inference}
\description{
vmpaR implements VMPA, a context-matched protein activity scoring
workflow for transcriptomic data.
}
\details{
Use \code{\link[=vmpa]{vmpa()}} to run VMPA for a selected cancer context with either
GSVA or FGSEA. For \code{algorithm = "gsva"}, the input is a gene-by-sample
expression matrix or data frame. For \code{algorithm = "fgsea"}, the input is
a named numeric vector of gene-level statistics or rankings.

Use \code{\link[=vmpa_gsc]{vmpa_gsc()}} to retrieve VMPA gene-set collections for a selected
context without running downstream scoring.
}
\section{Reference data}{

The context-specific VMPA reference subsets are available from Figshare at
\url{https://doi.org/10.6084/m9.figshare.32060643}.
}

\section{Example data}{

The package includes the example dataset \link{kebir_gb}, a small glioblastoma
\code{Biobase::ExpressionSet}. Expression values can be accessed with
\code{Biobase::exprs(kebir_gb)} and used as input for GSVA-based VMPA scoring.
}

\section{Citation}{

Please cite the VMPA paper when using vmpaR or VMPA-derived scores
in scientific work.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/landgrebe-a/vmpaR}
  \item \doi{10.6084/m9.figshare.32060643}
  \item Report bugs at \url{https://github.com/landgrebe-a/vmpaR/issues}
}

}
\author{
\strong{Maintainer}: Antonia Landgrebe \email{antonia.landgrebe@stud.uni-due.de}

Authors:
\itemize{
  \item Igor Cima
}

}
\keyword{internal}
