% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protivity_package.R
\docType{package}
\name{protivity-package}
\alias{protivity}
\alias{protivity-package}
\title{protivity: context-matched protein activity inference}
\description{
protivity implements COMPASS, a context-matched protein activity scoring
workflow for transcriptomic data.
}
\details{
Use \code{\link[=compass]{compass()}} to run COMPASS for a selected cancer context with either
GSVA or FGSEA. For \code{algorithm = "gsva"}, the input is a gene-by-sample
expression matrix or data frame. For \code{algorithm = "fgsea"}, the input is
a named numeric vector of gene-level statistics or rankings.

Use \code{\link[=compass_gsc]{compass_gsc()}} to retrieve COMPASS gene-set collections for a selected
context without running downstream scoring.
}
\section{Example data}{

The package includes the example dataset \link{kebir_gb}, a small glioblastoma
\code{Biobase::ExpressionSet}. Expression values can be accessed with
\code{Biobase::exprs(kebir_gb)} and used as input for GSVA-based COMPASS scoring.
}

\section{Citation}{

Please cite the COMPASS paper when using protivity or COMPASS-derived scores
in scientific work.
}

\author{
\strong{Maintainer}: Antonia Landgrebe \email{antonia.landgrebe@stud.uni-due.de}

Authors:
\itemize{
  \item Igor Cima
}

}
\keyword{internal}
